backend/requirements.txt
fastapi
uvicorn
requests
openai
pydantic
python-dotenv
pypdf
python-multipart
